+++ src/main/java/org/example/final_project/model/ShopModel.java
@@ -1,7 +1,6 @@
 package org.example.final_project.model;
 
 import lombok.*;
-import org.springframework.web.multipart.MultipartFile;
 @Getter
 @Setter
 @NoArgsConstructor
@@ -10,7 +9,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class ShopModel {
 
     private String shop_name;
-    private int shop_address;
+    private Long shop_address;
     private String shop_address_detail;
     private String phone;
 }
